review updates day 1
