+++ src/main/java/com/assignment/xml/vo/FindNos.java
@@ -4,7 +4,14 @@ import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 
+import java.util.List;
+
 import javax.xml.bind.annotation.XmlAccessType;
+
+/**
+ * @author keerthana
+ * this class represents the FindNo fields
+ */
 
 @XmlRootElement(name = "FIND_NO")
 @XmlAccessorType(XmlAccessType.FIELD)
@@ -16,15 +23,19 @@ public class FindNos {
 	
 	@XmlElement (name = "Description")
 	private String description;
+	
+	@XmlElement (name = "Options")
+	private List<Option> options;
 
 	public FindNos() {
 		super();
 	}
 
-	public FindNos(Integer id, String description) {
+	public FindNos(Integer id, String description, List<Option> options) {
 		super();
 		this.id = id;
 		this.description = description;
+		this.options = options;
 	}
 
 	public Integer getId() {
@@ -41,6 +52,14 @@ public class FindNos {
 
 	public void setDescription(String description) {
 		this.description = description;
+	}
+
+	public List<Option> getOptions() {
+		return options;
+	}
+
+	public void setOptions(List<Option> options) {
+		this.options = options;
 	}
