+++ src/main/java/com/assignment/xml/vo/KMat.java
@@ -4,7 +4,15 @@ import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 
+import java.util.List;
+
 import javax.xml.bind.annotation.XmlAccessType;
+
+
+/**
+ * @author keerthana
+ * this class represents the KMAT fields
+ */
 
 @XmlRootElement(name = "KMAT")
 @XmlAccessorType(XmlAccessType.FIELD)
@@ -16,11 +24,16 @@ public class KMat {
 	
 	@XmlElement (name = "MPN")
 	private String mpn;
+	
+	@XmlElement (name = "FIND_NOS")
+	private List<FindNos> findNos;
 
-	public KMat(String id, String mpn) {
+
+	public KMat(String id, String mpn, List<FindNos> findNos) {
 		super();
 		this.id = id;
 		this.mpn = mpn;
+		this.findNos = findNos;
 	}
 
 	public KMat() {
@@ -42,6 +55,15 @@ public class KMat {
 	public void setMpn(String mpn) {
 		this.mpn = mpn;
 	}
+
+	public List<FindNos> getFindNos() {
+		return findNos;
+	}
+
+	public void setFindNos(List<FindNos> findNos) {
+		this.findNos = findNos;
+	}
+	
 	
 	
 }
